+++ tests/testthat/test-category.R
@@ -61,3 +61,20 @@ test_that("climatology = T causes a list output with the time series category da
   expect_equal(ncol(cat$climatology), 4)
   expect_equal(nrow(cat$climatology), 900)
 })
+
+test_that("no detected events returns an empty dataframe and not an error", {
+  sst_WA_flat <- sst_WA
+  sst_WA_flat$temp <- 1
+  # sst_WA_flat$temp[4:6] <- 5
+  # ts_xy <- ts2clm(sst_WA_flat, climatologyPeriod = c("1983-01-01", "2012-12-31"))
+  res <- detect_event(ts2clm(sst_WA_flat, climatologyPeriod = c("1983-01-01", "2012-12-31")))
+  cat_event <- category(res, climatology = F)
+  cat_clim <- category(res, climatology = T)
+  expect_is(cat_event, "tbl_df")
+  expect_is(cat_clim, "list")
+  expect_is(cat_clim$climatology, "tbl_df")
+  expect_equal(nrow(cat_event), 0)
+  expect_equal(ncol(cat_event), 11)
+  expect_equal(nrow(cat_clim$climatology), 0)
+  expect_equal(ncol(cat_clim$climatology), 4)
+})